scripts/tools/vllm_inference/docker/requirements.txt
fire
azure-ai-ml>=1.14.0
azureml-core>=1.56.0
azure-identity
azure-keyvault-secrets
ipykernel